fixed warnings and added msvc projects for atsc_in

+++ src/utils/cache.c
@@ -992,7 +992,7 @@ Bool gf_cache_set_range(const DownloadedCacheEntry entry, u64 size, u64 start_ra
 	if (!entry || !entry->memory_stored) return GF_FALSE;
 	entry->range_start = start_range;
 	entry->range_end = end_range;
-	entry->contentLength = size;
+	entry->contentLength = (u32) size;
 	return GF_TRUE;
 }
